cspm-api/requirements.txt
fastapi[standard]==0.128.0
sqlmodel[async]==0.0.31
python-dotenv==1.2.1
pydantic-settings==2.12.0
pre-commit==4.5.1
asyncpg==0.31.0
greenlet==3.3.0
